Implement Decision Making for Player and Dealer
Implemented the decision logic that is to be used by both the player and the dealer.

Player utilizes Basic Strategy rules.
Dealer utilizes dealer's default strategy of stand on 17 above.

Added more helper functions to Dealer, that are similar if not the same as the ones in Player.

Now both Player and Dealer make use of enum defined in definitions.py.