Rework Basic Strategy and Use PyDealer
Reworked basic strategy logic that player uses  to be in line with what is in the README for the project.

Updated both the Player and the Dealer class to make use of the PyDealer module for cards.

Both now expect a list of Card objects when computing their decision.

Helper functions have been modified to conform with PyDealer's API and created new helper functions where needed.