Rework hand total with Aces
Modified how the sum of the hand is calculated when Aces are present.

Now it considers the first Ace it sees as 11 when counting the higher total, and every subsequent Ace gets considered as 1.

For the lower total, it still considers all Aces as 1.

Also added a new enum for Actions called NONE. The player's default decision is set to this enum and an error is raised if the decision does not get changed from the default value of NONE.